requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2020.6.24=0
certifi=2020.6.20=py38_0
openssl=1.1.1g=he774522_0
pillow=7.2.0=pypi_0
pip=20.2.2=py38_0
pymupdf=1.17.5=pypi_0
python=3.8.5=he1778fa_0
reportlab=3.5.47=pypi_0
setuptools=49.2.0=py38_0
sqlite=3.32.3=h2a8f88b_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.34.2=py38_0
wincertstore=0.2=py38_0
zlib=1.2.11=h62dcd97_4
